+++ src/Endpoints/BorgRepositories.php
@@ -2,8 +2,10 @@
 
 namespace Vdhicts\Cyberfusion\ClusterApi\Endpoints;
 
+use DateTimeInterface;
 use Vdhicts\Cyberfusion\ClusterApi\Exceptions\RequestException;
 use Vdhicts\Cyberfusion\ClusterApi\Models\BorgRepository;
+use Vdhicts\Cyberfusion\ClusterApi\Models\BorgRepositoryUsage;
 use Vdhicts\Cyberfusion\ClusterApi\Request;
 use Vdhicts\Cyberfusion\ClusterApi\Response;
 use Vdhicts\Cyberfusion\ClusterApi\Support\ListFilter;
@@ -66,6 +68,42 @@ class BorgRepositories extends Endpoint
     }
 
     /**
+     * @param int $id
+     * @param DateTimeInterface|null $from
+     * @return Response
+     * @throws RequestException
+     */
+    public function usages(int $id, DateTimeInterface $from = null): Response
+    {
+        $url = $this->applyOptionalQueryParameters(
+            sprintf('borg-repositories/usages/%d', $id),
+            ['from_timestamp_date' => $from]
+        );
+
+        $request = (new Request())
+            ->setMethod(Request::METHOD_GET)
+            ->setUrl($url);
+
+        $response = $this
+            ->client
+            ->request($request);
+        if (! $response->isSuccess()) {
+            return $response;
+        }
+
+        return $response->setData([
+            'borgRepositoryUsage' => count($response->getData()) !== 0
+                ? array_map(
+                    function (array $data) {
+                        return (new BorgRepositoryUsage())->fromArray($data);
+                    },
+                    $response->getData()
+                )
+                : null,
+        ]);
+    }
+
+    /**
      * @param BorgRepository $borgRepository
      * @return Response
      * @throws RequestException
@@ -75,11 +113,8 @@ class BorgRepositories extends Endpoint
         $this->validateRequired($borgRepository, 'create', [
             'name',
             'passphrase',
-            'keep_hourly',
-            'keep_daily',
-            'keep_weekly',
-            'keep_monthly',
-            'keep_yearly',
+            'remote_url',
+            'ssh_key_id',
             'unix_user_id',
         ]);
 
@@ -94,7 +129,56 @@ class BorgRepositories extends Endpoint
                 'keep_weekly',
                 'keep_monthly',
                 'keep_yearly',
+                'remote_url',
+                'ssh_key_id',
                 'unix_user_id',
+            ]));
+
+        $response = $this
+            ->client
+            ->request($request);
+        if (! $response->isSuccess()) {
+            return $response;
+        }
+
+        return $response->setData([
+            'borgRepository' => (new BorgRepository())->fromArray($response->getData()),
+        ]);
+    }
+
+    /**
+     * @param BorgRepository $borgRepository
+     * @return Response
+     * @throws RequestException
+     */
+    public function update(BorgRepository $borgRepository): Response
+    {
+        $this->validateRequired($borgRepository, 'update', [
+            'name',
+            'passphrase',
+            'remote_url',
+            'ssh_key_id',
+            'unix_user_id',
+            'id',
+            'cluster_id',
+        ]);
+
+        $request = (new Request())
+            ->setMethod(Request::METHOD_PUT)
+            ->setUrl(sprintf('borg-repositories/%d', $borgRepository->getId()))
+            ->setBody($this->filterFields($borgRepository->toArray(), [
+                'name',
+                'passphrase',
+                'keep_hourly',
+                'keep_daily',
+                'keep_weekly',
+                'keep_monthly',
+                'keep_yearly',
+                'remote_url',
+                'ssh_key_id',
+                'unix_user_id',
+                'id',
+                'cluster_id',
             ]));
 
         $response = $this