+++ src/Enums/ClusterGroup.php
@@ -7,5 +7,6 @@ class ClusterGroup
     public const WEB = 'Web';
     public const EMAIL = 'Mail';
     public const DATABASE = 'Database';
-    public const BORG = 'Borg';
+    public const BORG_CLIENT = 'Borg Client';
+    public const BORG_SERVER = 'Borg Server';
 }